+++ include/core/mock_core.h
@@ -11,7 +11,10 @@ public:
   mock_core();
 
   void start();
+
   void add_stream(std::shared_ptr<streams_t> stream);
+  void reload_rules();
+
   index_interface & idx();
   scheduler_interface & sched();
   external_interface & externals();