+++ include/scheduler/scheduler.h
@@ -9,12 +9,14 @@
 using task_fn_t = std::function<void()>;
 using remove_task_fn_t = std::function<void()>;
 using remove_task_future_t = std::future<remove_task_fn_t>;
+using remove_ns_tasks_future_t = std::future<bool>;
 
 class scheduler_interface {
 public:
   virtual remove_task_future_t add_periodic_task(task_fn_t task,
                                                   float interval) = 0;
   virtual remove_task_future_t add_once_task(task_fn_t task, float dt) = 0;
+  virtual remove_ns_tasks_future_t remove_ns_tasks(const std::string nm) = 0;
   virtual time_t unix_time() = 0;
   virtual void set_time(const time_t t) = 0;
   virtual void clear() = 0;