+++ src/async/real_async_loop.cpp
@@ -1,5 +1,6 @@
 #include <netinet/in.h>
 #include <glog/logging.h>
+#include <core/core.h>
 #include <async/real_async_loop.h>
 
 namespace {
@@ -16,6 +17,11 @@ int ev_mode(const async_fd::mode & initial_mode) {
     mode |= EV_WRITE | EV_READ;
   }
   return  mode;
+}
+
+unsigned long now() {
+   return  std::chrono::system_clock::now().time_since_epoch()
+           / std::chrono::milliseconds(1);
 }
 
 };
@@ -84,10 +90,16 @@ async_loop& real_async_fd::loop() {
   return async_loop_;
 }
 
-real_async_loop::real_async_loop() : async_loop(), stop_(false)
+real_async_loop::real_async_loop()
+  :
+    async_loop(),
+    stop_(false),
+    next_timer_id_(0)
 {
   async_.set(loop_);
   async_.set<real_async_loop, &real_async_loop::async_callback>(this);
+  timer_.set(loop_);
+  timer_.set<real_async_loop, &real_async_loop::timer_callback>(this);
 }
 
 void real_async_loop::set_id(size_t id) {
@@ -149,100 +161,161 @@ ev::dynamic_loop & real_async_loop::loop() {
   return loop_;
 }
 
-timer_id_t real_async_loop::add_timer(const timer_cb_fn_t task,
-                                      const bool once, const float t)
+timer_id_t real_async_loop::add_task(const std::string lib_namespace,
+                                     const timer_cb_fn_t task,
+                                     const bool once, const float t)
 {
-  auto timer_id = next_timer_id_++;
+  const auto timer_id = next_timer_id_++;
+  const unsigned long current_time = now();
+  const unsigned long t_ms = static_cast<unsigned long>(t * 1000);
+  const sched_task_t sched_task{lib_namespace,
+                                [=](){ task(id_); },
+                                {id_, timer_id},
+                                once ? 0 : t_ms,
+                                current_time + t_ms};
+  tasks_.push(sched_task);
 
-  auto  fn = [=]()
-  {
-    task(id_);
-
-    if (once) { // Move this to static fun
-      auto it = timers_.find(timer_id);
-      CHECK(it != timers_.end()) << "timer not found";
-      timers_.erase(it);
-    }
-  };
-
-  timer_ctx_t ctx{{std::make_shared<ev::timer>(loop_)},
-                   std::make_shared<task_cb_fn_t>(fn),
-                   *this,
-                   {id_, timer_id},
-                   once};
-
-  auto it = timers_.insert({timer_id, std::make_shared<timer_ctx_t>(ctx)});
-  auto ev_timer = it.first->second->timer;
-
-  auto ctx_ptr = static_cast<void*>(it.first->second.get());
-  ev_timer->set<timer_callback>(ctx_ptr);
-
-  ev_timer->start(t, once ? 0 : t);
+  sched_next_task();
 
   return {id_, timer_id};
 }
 
-void real_async_loop::timer_callback(ev::timer & timer, int) {
+void real_async_loop::sched_next_task() {
+  VLOG(3) << "sched_next_task";
 
+  if (tasks_.empty()) {
+    VLOG(3) << "task queue is empty";
+    timer_.stop();
+    return;
+  }
+
+  unsigned long next_time = tasks_.top().time_ms;
+  unsigned long current_time = now();
+
+  VLOG(3) << "next_time: " <<  next_time << " current_time: " << current_time;
+
+  float next;
+  if (current_time > next_time) {
+    next = 0;
+  } else {
+    next = (next_time - current_time) / 1000.0;
+  }
+
+  VLOG(3) << "next: " << next;
+
+  timer_.set<real_async_loop, &real_async_loop::timer_callback>(this);
+
+  timer_.start(next);
+}
+
+void real_async_loop::timer_callback(ev::timer &, int) {
   VLOG(3) << "timer_callback()";
 
-  auto ctx_ptr = static_cast<timer_ctx_t*>(timer.data);
+  while (!tasks_.empty()) {
 
-  (*ctx_ptr->task)();
+    auto task = tasks_.top();
 
-  if (ctx_ptr->once) {
-    ctx_ptr->async_loop.remove_task(ctx_ptr->id);
+    if (task.time_ms > now()) {
+      break;
+    }
+
+    task.task();
+
+    // The task we just run may have removed itself
+    if (tasks_.empty()) {
+      continue;
+    }
+
+    if (tasks_.top().id.timer_id != task.id.timer_id) {
+      continue;
+    }
+
+    tasks_.pop();
+
+    if (task.interval_ms > 0) {
+      task.time_ms = now() + task.interval_ms;
+      tasks_.push(task);
+    }
+
   }
+
+  sched_next_task();
 
 }
 
-timer_id_t real_async_loop::add_once_task(const timer_cb_fn_t task,
+timer_id_t real_async_loop::add_once_task(const std::string lib_namespace,
+                                          const timer_cb_fn_t task,
                                           const float t) {
   VLOG(3) << "add_once_task() t: " << t;
 
-  return add_timer(task, true, t);
+  return add_task(lib_namespace, task, true, t);
 
 }
 
-timer_id_t real_async_loop::add_periodic_task(const timer_cb_fn_t task,
+timer_id_t real_async_loop::add_periodic_task(const std::string lib_namespace,
+                                              const timer_cb_fn_t task,
                                               const float t) {
   VLOG(3) << "add_periodic_task() t: " << t;
 
-  return add_timer(task, false, t);
+  return add_task(lib_namespace, task, false, t);
 
 }
 
-void real_async_loop::set_task_interval(const timer_id_t id, const float t) {
+void real_async_loop::set_task_interval(const timer_id_t, const float t) {
 
+  // TODO
   VLOG(3) << "set_task_interval() t: " << t;
   return;
 
-  auto it = timers_.find(id.timer_id);
-  CHECK(it != timers_.end()) << "timer not found";
-
-  auto & ev_timer = it->second->timer;
-  ev_timer->start(t, t);
-
 }
 
-bool real_async_loop::remove_task(const timer_id_t id) {
 
+bool real_async_loop::remove_task(
+    const std::function<bool(const sched_task_t &)> & predicate)
+{
   VLOG(3) << "remove_task()";
 
-  auto it = timers_.find(id.timer_id);
+  std::priority_queue<sched_task_t,
+                      std::vector<sched_task_t>,
+                      sched_task_cmp> new_tasks;
 
-  if (it != timers_.end()) {
+  bool removed = false;
+  while (!tasks_.empty()) {
+    auto task = tasks_.top();
+    tasks_.pop();
 
-   it->second->timer->stop();
-   timers_.erase(it);
+    if (predicate(task)) {
+      removed = false;
+      continue;
+    }
 
-  return true;
-
-  } else {
-
-    VLOG(3) << "timer not found";
-    return false;
+    new_tasks.push(task);
   }
+
+  tasks_ = new_tasks;
+
+  return removed;
+}
+
+bool real_async_loop::remove_task(const timer_id_t timer_id) {
+  return remove_task(
+      [=](const sched_task_t & task)
+      {
+        return (task.id.timer_id == timer_id.timer_id);
+      }
+  );
+}
+
+void real_async_loop::remove_task_lib_namespace(const std::string lib_namespace)
+{
+  VLOG(3) << "remove task for " << lib_namespace;
+
+  remove_task(
+      [=](const sched_task_t & task)
+      {
+        return (task.lib_namespace == lib_namespace);
+      }
+  );
 
 }
 
@@ -343,9 +416,11 @@ real_main_async_loop::real_main_async_loop()
   sigterm_.set<real_main_async_loop, &real_main_async_loop::signal_cb>(this);
   sigterm_.start(SIGTERM);
 
+  sigterm_.set<real_main_async_loop, &real_main_async_loop::signal_cb>(this);
+  sigterm_.start(SIGHUP);
+
   async_.set<real_main_async_loop, &real_main_async_loop::async_cb>(this);
   async_.start();
-
 }
 
 void real_main_async_loop::start() {
@@ -357,8 +432,16 @@ void real_main_async_loop::add_tcp_listen_fd(const int fd,
   listen_ios_.emplace_back(std::make_shared<listen_io>(fd, on_new_client));
 }
 
-void real_main_async_loop::signal_cb(ev::sig &, int) {
-  default_loop_.break_loop();
+void real_main_async_loop::signal_cb(ev::sig & sig, int) {
+
+  VLOG(3) << "signal_cb() " << sig.signum;
+
+  if (sig.signum == SIGHUP) {
+    g_core->reload_rules();
+  } else {
+    default_loop_.break_loop();
+  }
+
 }
 
 void real_main_async_loop::add_periodic_task(task_cb_fn_t task,