+++ tests/rules_common_test_case.hpp
@@ -19,12 +19,12 @@ TEST(critical_above_test_case, test)
 
   e.set_metric_d(1);
   push_event(cabove, e);
-  ASSERT_EQ(1, v.size());
+  ASSERT_EQ(1u, v.size());
   ASSERT_EQ("ok", v[0].state());
 
   e.set_metric_d(6);
   push_event(cabove, e);
-  ASSERT_EQ(2, v.size());
+  ASSERT_EQ(2u, v.size());
   ASSERT_EQ("critical", v[1].state());
 }
 
@@ -58,11 +58,11 @@ TEST(stable_metric_above_test_case, test)
   e.set_metric_d(1);
   e.set_time(1);
   push_event(td_above, e);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_time(6);
   push_event(td_above, e);
-  ASSERT_EQ(2, v.size());
+  ASSERT_EQ(2u, v.size());
   ASSERT_EQ("ok", v[0].state());
   ASSERT_EQ("ok", v[1].state());
   v.clear();
@@ -70,15 +70,15 @@ TEST(stable_metric_above_test_case, test)
   e.set_metric_d(6);
   e.set_time(10);
   push_event(td_above, e);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_time(12);
   push_event(td_above, e);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_time(15);
   push_event(td_above, e);
-  ASSERT_EQ(3, v.size());
+  ASSERT_EQ(3u, v.size());
   ASSERT_EQ("critical", v[0].state());
   ASSERT_EQ("critical", v[1].state());
   ASSERT_EQ("critical", v[2].state());
@@ -97,11 +97,11 @@ TEST(stable_metric_under_test_case, test)
   e.set_metric_d(1);
   e.set_time(1);
   push_event(td_under, e);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_time(6);
   push_event(td_under, e);
-  ASSERT_EQ(2, v.size());
+  ASSERT_EQ(2u, v.size());
   ASSERT_EQ("critical", v[0].state());
   ASSERT_EQ("critical", v[1].state());
   v.clear();
@@ -109,15 +109,15 @@ TEST(stable_metric_under_test_case, test)
   e.set_metric_d(6);
   e.set_time(10);
   push_event(td_under, e);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_time(12);
   push_event(td_under, e);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_time(15);
   push_event(td_under, e);
-  ASSERT_EQ(3, v.size());
+  ASSERT_EQ(3u, v.size());
   ASSERT_EQ("ok", v[0].state());
   ASSERT_EQ("ok", v[1].state());
   ASSERT_EQ("ok", v[2].state());
@@ -145,19 +145,19 @@ TEST(agg_sum_trigger_above_test_case,test)
   e2.set_time(1);
   push_event(agg,  e1);
   push_event(agg,  e2);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e1.set_time(6);
   e2.set_time(6);
   push_event(agg,  e1);
   push_event(agg,  e2);
 
-  ASSERT_EQ(4, v.size());
+  ASSERT_EQ(4u, v.size());
   v.clear();
 
   e1.set_time(7);
   push_event(agg,  e1);
-  ASSERT_EQ(1, v.size());
+  ASSERT_EQ(1u, v.size());
   ASSERT_EQ("ok", v[0].state());
   ASSERT_EQ(2, v[0].metric_d());
   v.clear();
@@ -165,15 +165,15 @@ TEST(agg_sum_trigger_above_test_case,test)
   e2.set_time(10);
   e2.set_metric_d(5);
   push_event(agg,  e2);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e1.set_time(14);
   push_event(agg,  e1);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e2.set_time(16);
   push_event(agg,  e2);
-  ASSERT_EQ(3, v.size());
+  ASSERT_EQ(3u, v.size());
   ASSERT_EQ("critical", v[0].state());
   ASSERT_EQ(6, v[0].metric_d());
 }
@@ -195,17 +195,17 @@ TEST(max_critical_hosts_test_case,test)
     push_event(max, events[i]);
   }
 
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   events[0].set_state("critical");
   push_event(max, events[0]);
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   events[1].set_state("critical");
   events[2].set_state("critical");
   push_event(max, events[1]);
   push_event(max, events[2]);
-  ASSERT_EQ(1, v.size());
+  ASSERT_EQ(1u, v.size());
   ASSERT_EQ("critical", v[0].state());
 
   v.clear();
@@ -216,7 +216,7 @@ TEST(max_critical_hosts_test_case,test)
   push_event(max, events[0]);
   push_event(max, events[1]);
   push_event(max, events[2]);
-  ASSERT_EQ(1, v.size());
+  ASSERT_EQ(1u, v.size());
   ASSERT_EQ("ok", v[0].state());
 
 }
@@ -255,7 +255,7 @@ TEST(per_host_ratio_test_case,test)
 
   push_event(ratio, e);
 
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_service("a");
   e.set_metric_d(1);
@@ -275,7 +275,7 @@ TEST(per_host_ratio_test_case,test)
   e.set_metric_d(1);
   push_event(ratio, e);
 
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_service("a");
   e.set_metric_d(1);
@@ -286,7 +286,7 @@ TEST(per_host_ratio_test_case,test)
   e.set_metric_d(1);
   push_event(ratio, e);
 
-  ASSERT_EQ(3, v.size());
+  ASSERT_EQ(3u, v.size());
 
   ASSERT_EQ("critical", v[0].state());
   ASSERT_EQ("critical", v[1].state());
@@ -303,7 +303,7 @@ TEST(per_host_ratio_test_case,test)
   e.set_metric_d(5);
   push_event(ratio, e);
 
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   e.set_service("a");
   e.set_metric_d(1);
@@ -314,7 +314,7 @@ TEST(per_host_ratio_test_case,test)
   e.set_metric_d(5);
   push_event(ratio, e);
 
-  ASSERT_EQ(2, v.size());
+  ASSERT_EQ(2u, v.size());
   ASSERT_EQ("ok", v[0].state());
   ASSERT_EQ("ok", v[1].state());
 
@@ -336,14 +336,14 @@ TEST(stable_event_stream_test_case,test)
     push_event(stable, e);
   }
 
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
   for (int i = 0; i < 3; i++) {
     e.set_state("critical");
     push_event(stable, e);
   }
 
-  ASSERT_EQ(1, v.size());
+  ASSERT_EQ(1u, v.size());
 
   v.clear();
 
@@ -352,7 +352,7 @@ TEST(stable_event_stream_test_case,test)
     push_event(stable, e);
   }
 
-  ASSERT_EQ(0, v.size());
+  ASSERT_EQ(0u, v.size());
 
 
 }